Fix accessing unallocated memory
In `mul()`, it assumes that the multiplication causes a carry, so
`bitoa()`, which calls `mul()`, needs to allocate one btye of extra
memory.
For example, "9" * "1" = "9", the len of the result is 1. However, in
`mul()`, due to the assumation that the maximum length of the result of
multiplication of 1 digit by 1 digit is 2 digits, `mul()` needs 2 bytes
to work correctly.