New: `bigint_mul_bigint_school`
Add three functions in the source file: `bigint_mul_bigint_school`,
`split`, `combine`.
`split` splits one bigint_chunk into two, and `combine` merges them
back. In original, assuming the number of bits of `bigint_chunk` is 8,
the data of a number is stored as `0b01100110 0b01111000` of which the
hightest bits is preserved for doing additions. After calling `split`,
the format of the data becomes
`0b00000110 0b00000110 0b00001100 0b00000011`. The 14 bits of data is
separated into 4 chunks, each of which keeps at most 4 bits, half of 8.
And as mentioned, `combine` can convert this format back to the original
format.