Add property `size` to `Bigint`
- Rename the original property `len` to `capacity`
- Rewrite the tests
- Add `trim_birs_by_one()` to correct the value of `bits`

`capacity` is the property of the number of the bytes allocated. `size`
is the number of bytes keeping the data.

The `capacity` of the Bigint of which value is 0 can be zero, and the
`chunk` of the Bigint may be NULL, and the `size` and `bits` must be
zero, and `sign` must be one.